Add draw_board header
